+++ adc.h
@@ -11,6 +11,6 @@ void adc_init(void);
  */
 int adc_read(char channel);
 
-int adc_read_voltage(char channel);
+float adc_read_voltage(char channel);
 
 #endif // ADC_H
